fastapi-backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
scikit-learn
openai-whisper
torch
torchaudio

pytesseract
openai-whisper
moviepy
SpeechRecognition
requests
python-docx
PyPDF2
docx2txt
beautifulsoup4
pillow
python-multipart